# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: O-RAN × Nephio RAG API
  description: |
    Developer-friendly REST API for O-RAN and Nephio knowledge retrieval using advanced RAG (Retrieval-Augmented Generation) technology.

    This API provides intelligent Q&A capabilities for telecommunications and cloud-native network functions,
    specialized for O-RAN (Open Radio Access Network) and Nephio platform documentation.

    ## Features
    - 🤖 **AI-Powered Q&A**: Advanced natural language processing for technical queries
    - 📚 **Official Documentation**: Automated ingestion from O-RAN Alliance and Nephio sources
    - 🔍 **Semantic Search**: High-precision vector-based document retrieval
    - ⚡ **Multiple Deployment Modes**: Browser automation, mock testing, and local inference
    - 📊 **Comprehensive Monitoring**: Built-in metrics, logging, and health checks
    - 🔒 **Enterprise Security**: Authentication, rate limiting, and audit trails

    ## Getting Started
    1. Set up your API mode (browser/mock) in environment variables
    2. Initialize the vector database with official documentation
    3. Start making queries to get intelligent answers about O-RAN and Nephio

    ## Rate Limiting
    - Query endpoints: 10 requests per minute per IP
    - Search endpoints: 20 requests per minute per IP
    - Bulk operations: 2 requests per minute per IP

    ## Authentication
    API key authentication is required for production deployments.
...
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/v1/documents:
    get:
      tags: [Documents]
      summary: List Documents
      description: Get a list of all documents in the vector database
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: source_type
          in: query
          description: Filter by source type
          schema:
            type: string
            enum: [nephio, oran_sc]
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: Document list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'

    post:
      tags: [Documents]
      summary: Add Document
      description: Add a new document source to the system
      operationId: addDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentRequest'
            example:
              url: "https://docs.nephio.org/docs/new-feature/"
              source_type: "nephio"
              description: "New Nephio feature documentation"
              priority: 2
              enabled: true
      responses:
        '201':
          description: Document added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid document request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/{document_id}:
    get:
      tags: [Documents]
      summary: Get Document
      description: Get details of a specific document
      operationId: getDocument
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Documents]
      summary: Update Document
      description: Update an existing document
      operationId: updateDocument
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Documents]
      summary: Delete Document
      description: Remove a document from the system
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, message, timestamp]
      properties:
        success:
          type: boolean
          example: false
          description: Always false for errors
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Query cannot be empty"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    DocumentRequest:
      type: object
      required: [url, source_type, description]
      properties:
        url:
          type: string
          format: uri
          description: Document URL
        source_type:
          type: string
          enum: [nephio, oran_sc]
          description: Source type
        description:
          type: string
          description: Document description
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
          description: Priority level (1=highest, 5=lowest)
        enabled:
          type: boolean
          default: true
          description: Whether document is enabled
    DocumentResponse:
      type: object
      required: [id, url, source_type, description, priority, enabled, status]
      properties:
        id:
          type: string
          description: Document ID
        url:
          type: string
          format: uri
          description: Document URL
        source_type:
          type: string
          description: Source type
        description:
          type: string
          description: Document description
        priority:
          type: integer
          description: Priority level
        enabled:
          type: boolean
          description: Whether document is enabled
        last_updated:
          type: string
          format: date-time
          description: Last update time
        content_length:
          type: integer
          description: Content length in characters
        status:
          type: string
          enum: [active, processing, error, disabled]
          description: Document status
    DocumentListResponse:
      type: object
      required: [documents, total, enabled, by_source_type]
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
          description: List of documents
        total:
          type: integer
          description: Total number of documents
        enabled:
          type: integer
          description: Number of enabled documents
        by_source_type:
          type: object
          additionalProperties:
            type: integer
          description: Count by source type
...